+++ src/py/flwr/server/connection/grpc_rere/grpc_rere_driver_connection.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Flower gRPC Driver."""
+"""Flower gRPC-rere DriverConnection."""
 
 import time
 import warnings
@@ -45,7 +45,7 @@ from flwr.proto.serverappio_pb2 import (  # pylint: disable=E0611
 from flwr.proto.serverappio_pb2_grpc import ServerAppIoStub  # pylint: disable=E0611
 from flwr.proto.task_pb2 import TaskIns  # pylint: disable=E0611
 
-from .driver import DriverConnection
+from ..driver_connection import DriverConnection
 
 ERROR_MESSAGE_DRIVER_NOT_CONNECTED = """
 [Driver] Error: Not connected.
@@ -55,8 +55,8 @@ Call `connect()` on the `GrpcDriverStub` instance before calling any of the othe
 """
 
 
-class GrpcDriver(DriverConnection):
-    """`GrpcDriver` provides an interface to the ServerAppIo API.
+class GrpcRereDriverConnection(DriverConnection):
+    """Network-based request-response driver connection via ServerAppIo API.
 
     Parameters
     ----------
